backend/requirements.txt
alembic==1.8.1
blinker==1.5
certifi==2022.9.24
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
config==0.5.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
flask-buzz==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
flask-praetorian==1.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
passlib==1.7.4
pendulum==2.1.2
py-buzz==2.1.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytzdata==2020.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.10.0
